Combine urls/views and sort templates into folders
