+++ apps/authentication/templates/partials/set_country_flag.html
@@ -1,0 +1,1 @@
+<img src="https://flagcdn.com/h240/{{countryid}}.png" class="mb-4 w-24 h-16 object-cover mx-auto">
