--- apps/authentication/templates/set_country_flag.html
@@ -1,1 +1,0 @@
-<img src="https://flagcdn.com/h240/{{countryid}}.png" class="mb-4 w-24 h-16 object-cover mx-auto">
