+++ apps/currencies/templates/partials/result.html
@@ -1,0 +1,3 @@
+<h1 class="text-xl font-normal font-serif text-black">
+    {{amount}} {{start_currency}} = {{conversion}} {{result_currency}}
+</h1>
