--- apps/currencies/templates/result.html
@@ -1,3 +1,0 @@
-<h1 class="text-xl font-normal font-serif text-black">
-    {{amount}} {{start_currency}} = {{conversion}} {{result_currency}}
-</h1>
