--- apps/myapp/templates/main/index.html
@@ -1,33 +1,0 @@
-{% extends "base.html" %}
-{% load static %}
-{% block content %}
-  
-<body class="bg-gray-50">
-    <div class="flex justify-center">
-        {% if user.is_authenticated %}
-            <a hx-get="/user/profile" hx-target="body" hx-push-url="true" class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded m-4 hover:cursor-pointer">
-                Profile
-            </a>
-            <a hx-get="/user/signout" hx-target="body" hx-push-url="true" class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded m-4 hover:cursor-pointer">
-                Sign Out
-            </a>
-            <a hx-get="{{ user.username }}/mytrips" hx-target="body" hx-push-url="true" class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded m-4 hover:cursor-pointer">
-                My Trips
-            </a>
-        {% else %}
-            <a hx-get="/user/register" hx-target="body" hx-push-url="true" class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded m-4 hover:cursor-pointer">
-                Register
-            </a>
-            <a hx-get="{% url 'signin' %}" hx-target="body" hx-push-url="true" class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded m-4 hover:cursor-pointer">
-                Sign In
-            </a>
-        {% endif %}
-    </div>
-    <br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br>
-    <br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br>
-    <button _="on click go to the top of the body smoothly">
-            Take A Trip
-    </button>
-</body>
-
-{% endblock %}
