--- apps/myapp/urls_main.py
@@ -1,8 +1,0 @@
-from django.urls import include, path
-from . import views_main as views
-
-urlpatterns = [
-    path('', views.home, name='home'),
-    path('<str:username>/mytrips', views.mytrips, name='mytrips'),
-    path('<str:username>/trip/<int:trip_id>', views.configtrip, name='configtrip')
-]
