--- apps/myapp/urls_partials.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from . import views_partials as views
-
-urlpatterns = [
-    path('cities', views.cities, name='cities'),
-]
